move temp_env_var to py_functional
