+++ verl/utils/py_functional.py
@@ -20,6 +20,7 @@ import multiprocessing
 import os
 import queue  # Import the queue module for exception type hint
 import signal
+from contextlib import contextmanager
 from functools import wraps
 from types import SimpleNamespace
 from typing import Any, Callable, Dict, Iterator, Optional, Tuple
@@ -270,6 +271,37 @@ class DynamicEnum(metaclass=DynamicEnumMeta):
     @classmethod
     def from_name(cls, name: str) -> Optional["DynamicEnum"]:
         return cls._registry.get(name.upper())
+
+
+@contextmanager
+def temp_env_var(key: str, value: str):
+    """Context manager for temporarily setting an environment variable.
+
+    This context manager ensures that environment variables are properly set and restored,
+    even if an exception occurs during the execution of the code block.
+
+    Args:
+        key: Environment variable name to set
+        value: Value to set the environment variable to
+
+    Yields:
+        None
+
+    Example:
+        >>> with temp_env_var("MY_VAR", "test_value"):
+        ...     # MY_VAR is set to "test_value"
+        ...     do_something()
+        ... # MY_VAR is restored to its original value or removed if it didn't exist
+    """
+    original = os.environ.get(key)
+    os.environ[key] = value
+    try:
+        yield
+    finally:
+        if original is None:
+            os.environ.pop(key, None)
+        else:
+            os.environ[key] = original
 
 
 def convert_to_regular_types(obj):
